Add appointment and comment models

+++ flutter_app/lib/models/site_detail_response.dart
@@ -1,7 +1,5 @@
-import 'dart:ffi';
-
 class SiteDetailResponse {
-  late Long id;
+  late int id;
   late String name;
   late String description;
   late String address;
